Final Update on Netflix Clone
HI!!! Mr Xola !!!I finally got the Search bar to work, and just formated the whole document i only worked on landing page as the login page was not Required for this project and wanted to focus on all the elements of the home page Movies are   able to play using youtube Trailers. Thanks Xola hope you like the project!